build: haskell-build-system: Support packages w. multiple libraries
As discussed in #53655 [0], haskell-build-system.scm does not currently
support packages which contain multiple libraries, such as attoparsec
[1]. Such libraries are often used for internal or convenience libraries, as
detailed in the documentation [2].

The current state of affairs means that attoparsec and any other libraries
with similar structure cannot be built on guix, nor can any library or
executable that depends on such a library.

Instead of assuming that `runhaskell Setup.hs --gen-pkg-config=...` always
outputs a single file, we detect whether the result was a directory. If so, we
need to handle each file independently in lexicographic order by calling the
new private function `install-from-config-file`.

Because `install-transitive-deps` can now be called multiple times for each
call to `register`, it has also been amended to support the files already
installed.

0: https://issues.guix.gnu.org/53655
1: https://hackage.haskell.org/package/attoparsec-0.14.4/attoparsec.cabal
2: https://downloads.haskell.org/cabal/Cabal-3.0.0.0/doc/users-guide/installing-packages.html#cmdoption-setup-register-gen-pkg-config